adjust styles, move puzzle pages under /play
