+++ src/lib/components/graphics/FeaturedPuzzle.svelte
@@ -1,4 +1,4 @@
-<svg width="36" height="36" viewBox="0 0 62 62" fill="none" xmlns="http://www.w3.org/2000/svg">
+<svg width="48" height="48" viewBox="0 0 62 62" fill="none" xmlns="http://www.w3.org/2000/svg">
 	<rect
 		x="1.5"
 		y="1.5"
