+++ src/routes/+page.svelte
@@ -1,8 +1,23 @@
-<script lang="ts">
-	import Logo from '$lib/components/graphics/Logo.svelte';
-
-	import { GridIcon, HelpCircleIcon, ListIcon, PlayIcon } from 'svelte-feather-icons';
+<script>
 	import LargeLink from '$lib/components/LargeLink.svelte';
+	import UnderlinedHeading from '$lib/components/UnderlinedHeading.svelte';
+	import FeaturedPuzzle from '$lib/components/graphics/FeaturedPuzzle.svelte';
+	import Logo from '$lib/components/graphics/Logo.svelte';
+	import { userStore } from '$lib/stores/user';
+	import {
+		ArrowLeftIcon,
+		AwardIcon,
+		BookmarkIcon,
+		CalendarIcon,
+		GridIcon,
+		HelpCircleIcon,
+		ListIcon
+	} from 'svelte-feather-icons';
+
+	export let data;
+
+	$: latestIsNew = data.latest && !$userStore.puzzles[data.latest.id];
+	$: latestPuzzle = data.latest?.data;
 </script>
 
 <svelte:head>
@@ -18,21 +33,58 @@
 	<Logo />
 </div>
 
-<h3>Play</h3>
 <div class="links">
-	<LargeLink href="/play" title="Play" highlightColor="var(--blue-light)">
-		<svelte:fragment slot="description">Daily puzzles, featured puzzles and more!</svelte:fragment>
+	<div class="row">
+		<LargeLink title="Daily" href="/play/daily" highlightColor="var(--blue-light)">
+			<svelte:fragment slot="description">Every day a new random puzzle!</svelte:fragment>
 
-		<PlayIcon size="48" strokeWidth={1.5} class="challenges" slot="icon" />
+			<span slot="icon" class="icon" style="color:var(--blue)">
+				<CalendarIcon size="48" strokeWidth={1} />
+			</span>
+		</LargeLink>
+
+		<LargeLink title="Random" href="/play/random">
+			<svelte:fragment slot="description">Infinite variations</svelte:fragment>
+
+			<span slot="icon" style="color:var(--red)">
+				<FeaturedPuzzle />
+			</span>
+		</LargeLink>
+	</div>
+
+	<br />
+
+	<LargeLink
+		href="/play/featured/latest"
+		title="Featured"
+		badge={latestIsNew ? 'New!' : null}
+		highlightColor="var(--orange-light)"
+	>
+		<svelte:fragment slot="description">
+			{#if latestPuzzle}
+				Spell “<b>{latestPuzzle.solution.toLowerCase()}</b>” within
+				<b>{latestPuzzle.maxMoves.bronze}</b> moves
+			{:else}
+				Solve the puzzle within the given moves
+			{/if}
+		</svelte:fragment>
+
+		<span slot="icon" style="display:contents;color:var(--orange);">
+			<BookmarkIcon size="48" strokeWidth={1} />
+		</span>
 	</LargeLink>
 
-	<LargeLink title="Campaigns" href="/campaign">
-		<GridIcon slot="icon" size="24" class="campaign" />
+	<LargeLink title="Puzzle archive" href="/archive">
+		<ListIcon slot="icon" />
 	</LargeLink>
 </div>
 
 <h3>Other</h3>
 <div class="links">
+	<LargeLink title="Campaigns" href="/campaign">
+		<GridIcon slot="icon" />
+	</LargeLink>
+
 	<LargeLink title="Tutorial" href="/tutorial">
 		<HelpCircleIcon slot="icon" />
 	</LargeLink>
@@ -46,19 +98,6 @@
 		margin-bottom: 50px;
 	}
 
-	:global(.challenges) {
-		opacity: 0.5;
-	}
-
-	:global(.campaign) {
-		opacity: 0.5;
-	}
-
-	.about {
-		width: fit-content;
-		color: var(--black);
-	}
-
 	h3 {
 		font-family: var(--font-body);
 		font-weight: 400;
@@ -67,11 +106,30 @@
 		text-align: center;
 	}
 
+	.about {
+		width: fit-content;
+		color: var(--black);
+	}
+
+	.icon {
+		display: flex;
+		margin-block: 0.5rem;
+	}
+
 	.links {
 		width: 100%;
 		display: flex;
 		flex-direction: column;
 		gap: 16px;
 		margin-bottom: 32px;
+	}
+
+	.row {
+		display: flex;
+		gap: 16px;
+
+		@media screen and (max-width: 500px) {
+			flex-direction: column;
+		}
 	}
 </style>
