+++ src/routes/play/+page.server.ts
@@ -1,16 +1,5 @@
-import { connectDB, getLatestPuzzle } from '$lib/services/database';
-import type { ServerLoad } from '@sveltejs/kit';
+import { redirect } from '@sveltejs/kit';
 
-export const load: ServerLoad = async ({ setHeaders }) => {
-	await connectDB();
-
-	const latest = await getLatestPuzzle();
-
-	setHeaders({
-		'Cache-Control': 'public, max-age=300' // cache for 5 minutes
-	});
-
-	return {
-		latest
-	};
+export const load = () => {
+	throw redirect(302, '/');
 };
