+++ src/routes/play/random/[seed]/+page.server.ts
@@ -5,7 +5,7 @@ import { error } from '@sveltejs/kit';
 export const load = async ({ params }) => {
 	try {
 		return {
-			puzzle: generatePuzzle(params.seed, { words, maxLength: 7 })
+			puzzle: generatePuzzle(params.seed, params.seed, { words, maxLength: 7 })
 		};
 	} catch (err: any) {
 		throw error(500, err?.message ?? 'unknown error');
